chore: isolates query in hooks
